Side menu hard loads when selected menu item renders Shesha setting forms #1936
