+++ shesha-functional-tests/adminportal/src/app/(main)/dynamic/[...path]/page.tsx
