+++ shesha-functional-tests/adminportal/src/app/(main)/settings/entity-configs/configurator/page.tsx
