--- shesha-functional-tests/adminportal/src/app/settings/layout.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import React from 'react';
-import { MainLayout } from '@shesha-io/reactjs';
-
-export default function Layout({
-    children,
-}: {
-    children: React.ReactNode;
-}) {
-    return (
-        <MainLayout noPadding>
-            {children}
-        </MainLayout>
-    );
-};
