+++ shesha-functional-tests/publicportal/src/app/(main)/settings/entity-configs/configurator/page.tsx
