+++ shesha-functional-tests/publicportal/src/app/(main)/settings/theme/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import React, { FC } from "react";
+import { ConfigurableThemePage } from "@shesha-io/reactjs";
+
+const Page: FC = () => {
+  return <ConfigurableThemePage />;
+};
+
+export default Page;
