+++ shesha-functional-tests/publicportal/src/app/(main)/shesha/settings/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import React, { FC } from "react";
+import { SettingsPage } from "@shesha-io/reactjs";
+
+const Page: FC = (props) => {
+  return <SettingsPage {...props} />;
+};
+
+export default Page;
