--- shesha-functional-tests/publicportal/src/app/settings/layout.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { PortalLayout } from "@shesha-io/pd-publicportal";
-import React from "react";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <PortalLayout>{children}</PortalLayout>;
-}
