+++ shesha-reactjs/src/app/(main)/settings/entity-configs/configurator/page.tsx
