+++ shesha-reactjs/src/app/(main)/settings/theme/page.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import React from 'react';
+import { ConfigurableThemePage } from '@/generic-pages/settings/dynamic-theme';
+import { PageWithLayout } from '@/interfaces';
+
+const Page: PageWithLayout<{}> = () => {
+  return <ConfigurableThemePage />;
+};
+
+export default Page;
