--- shesha-reactjs/src/app/dynamic/[...path]/layout.tsx
@@ -1,3 +1,0 @@
-import CommonLayout from "@/app/commonLayout";
-
-export default CommonLayout;
