--- shesha-reactjs/src/app/settings/layout.tsx
@@ -1,3 +1,0 @@
-import CommonLayout from "@/app/commonLayout";
-
-export default CommonLayout;
