--- shesha-reactjs/src/app/settings/theme/page.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-import React from 'react';
-import { ConfigurableThemePage } from '@/generic-pages/settings/dynamic-theme';
-import { PageWithLayout } from '@/interfaces';
-
-const Page: PageWithLayout<{}> = () => {
-  return <ConfigurableThemePage />;
-};
-
-export default Page;
