+++ shesha-starter/frontend/src/app/(main)/settings/entity-configs/configurator/page.tsx
