--- shesha-starter/frontend/src/app/settings/theme/page.tsx
@@ -1,10 +1,0 @@
-"use client";
-
-import React, { FC } from 'react';
-import { ConfigurableThemePage } from '@shesha-io/reactjs';
-
-const Page: FC = () => {
-  return <ConfigurableThemePage />;
-};
-
-export default Page;
