--- shesha-starter/frontend/src/app/shesha/layout.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { MainLayout } from '@shesha-io/reactjs';
-import React from 'react';
-
-export default function Layout({
-    children,
-}: {
-    children: React.ReactNode;
-}) {
-    return (
-        <MainLayout noPadding>
-            {children}
-        </MainLayout>
-    );
-};
