--- shesha-starter/frontend/src/app/shesha/settings/page.tsx
@@ -1,10 +1,0 @@
-"use client";
-
-import React, { FC } from 'react';
-import { SettingsPage } from '@shesha-io/reactjs';
-
-const Page: FC = (props) => {
-    return <SettingsPage {...props} />;
-};
-
-export default Page;
