+++ shesha-starter/publicportal/src/app/(main)/settings/entity-configs/configurator/page.tsx
